refactor: split providers/codex em módulos
