+++ src/providers/codex/mod.rs
@@ -2,730 +2,29 @@
 //! fallback session-log) normalizadas para `CodexRateLimits`. Port fiel de
 //! `src/providers/codex.ts`.
 
-use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
+mod app_server;
+mod normalize;
+mod session_log;
+mod types;
 
-use indexmap::IndexMap;
-use serde::{Deserialize, Serialize};
-use time::{Duration as TimeDuration, OffsetDateTime, UtcOffset};
-
-use super::iso_from_ms;
-use super::types::{
-    CodexQuotaExtra, ExtraUsage, ModelWindows, ProviderExtra, ProviderQuota, QuotaWindow,
+pub use app_server::{
+    normalize_appserver_rate_limits, run_appserver_protocol, CodexAppServerAccount,
+    CodexAppServerAccountReadResult, CodexAppServerCredits, CodexAppServerLimitBucket,
+    CodexAppServerRateLimitsReadResult, CodexAppServerWindow,
 };
-use crate::formatters::shared::{classify_window, normalize_plan, WindowKind};
+pub use normalize::build_codex_quota;
+pub use session_log::{extract_rate_limits, find_latest_session_file};
+pub use types::{CodexCredits, CodexLimitBucket, CodexRateLimits, CodexWindowRaw};
 
-// ---- Formato interno (snake_case = formato do session-log; é o Raw cacheável) ----
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CodexWindowRaw {
-    pub used_percent: f64,
-    pub window_minutes: i64,
-    pub resets_at: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CodexLimitBucket {
-    pub limit_id: String,
-    #[serde(default)]
-    pub limit_name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub primary: Option<CodexWindowRaw>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub secondary: Option<CodexWindowRaw>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CodexCredits {
-    pub has_credits: bool,
-    pub unlimited: bool,
-    pub balance: String,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct CodexRateLimits {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub primary: Option<CodexWindowRaw>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub secondary: Option<CodexWindowRaw>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub credits: Option<CodexCredits>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub plan_type: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub buckets: Option<IndexMap<String, CodexLimitBucket>>,
-}
-
-// ---- Helpers de conversão (puros) ----
-
-/// Unix SEGUNDOS → ISO UTC; None se `<= 0`.
-fn unix_to_iso(ts: i64) -> Option<String> {
-    if ts <= 0 {
-        None
-    } else {
-        Some(iso_from_ms((ts as u64) * 1000))
-    }
-}
-
-/// CodexWindowRaw → QuotaWindow (remaining = 100 - round(used_percent)).
-fn to_quota_window(raw: &CodexWindowRaw) -> QuotaWindow {
-    QuotaWindow {
-        remaining: 100.0 - raw.used_percent.round(),
-        resets_at: unix_to_iso(raw.resets_at),
-        window_minutes: Some(raw.window_minutes),
-        used: None,
-        severity: None,
-    }
-}
-
-/// `limit_name` (não-vazio) ou `limit_id`; `[_-]+`→espaço; titlecase por palavra; vazio→"Codex".
-fn format_bucket_label(bucket: &CodexLimitBucket) -> String {
-    let raw = bucket
-        .limit_name
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-        .unwrap_or(&bucket.limit_id);
-    let normalized: String = raw
-        .chars()
-        .map(|c| if c == '_' || c == '-' { ' ' } else { c })
-        .collect();
-    let normalized = normalized.trim();
-    if normalized.is_empty() {
-        return "Codex".to_string();
-    }
-    normalized
-        .split_whitespace()
-        .map(|word| {
-            let mut chars = word.chars();
-            match chars.next() {
-                Some(first) => first.to_uppercase().collect::<String>() + chars.as_str(),
-                None => String::new(),
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-/// Insere `qw` no kind certo de `windows` (fiveHour/sevenDay únicos; resto em other).
-fn place_window(windows: &mut ModelWindows, raw: &CodexWindowRaw) {
-    let qw = to_quota_window(raw);
-    match classify_window(Some(raw.window_minutes)) {
-        WindowKind::FiveHour if windows.five_hour.is_none() => windows.five_hour = Some(qw),
-        WindowKind::SevenDay if windows.seven_day.is_none() => windows.seven_day = Some(qw),
-        _ => windows.other.get_or_insert_with(Vec::new).push(qw),
-    }
-}
-
-/// Constrói `modelsDetailed` a partir dos buckets (ou fallback legacy primary/secondary).
-fn build_model_windows(limits: &CodexRateLimits) -> BTreeMap<String, ModelWindows> {
-    let mut models: BTreeMap<String, ModelWindows> = BTreeMap::new();
-
-    if let Some(buckets) = limits.buckets.as_ref().filter(|b| !b.is_empty()) {
-        for bucket in buckets.values() {
-            let mut windows = ModelWindows::default();
-            for raw in [bucket.primary.as_ref(), bucket.secondary.as_ref()]
-                .into_iter()
-                .flatten()
-            {
-                place_window(&mut windows, raw);
-            }
-            // Fallback de mapeamento quando as durações não classificam limpo.
-            if windows.five_hour.is_none() {
-                if let Some(p) = bucket.primary.as_ref() {
-                    windows.five_hour = Some(to_quota_window(p));
-                }
-            }
-            if windows.seven_day.is_none() {
-                if let Some(s) = bucket.secondary.as_ref() {
-                    windows.seven_day = Some(to_quota_window(s));
-                }
-            }
-            if windows.five_hour.is_none()
-                && windows.seven_day.is_none()
-                && windows.other.as_ref().map(Vec::is_empty).unwrap_or(true)
-            {
-                continue;
-            }
-            let base_name = format_bucket_label(bucket);
-            let mut name = base_name.clone();
-            let mut suffix = 2;
-            while models.contains_key(&name) {
-                name = format!("{base_name} ({suffix})");
-                suffix += 1;
-            }
-            models.insert(name, windows);
-        }
-    }
-
-    // Legacy: só primary/secondary, sem buckets.
-    if models.is_empty() && (limits.primary.is_some() || limits.secondary.is_some()) {
-        let mut windows = ModelWindows::default();
-        for raw in [limits.primary.as_ref(), limits.secondary.as_ref()]
-            .into_iter()
-            .flatten()
-        {
-            place_window(&mut windows, raw);
-        }
-        if windows.five_hour.is_none() {
-            if let Some(p) = limits.primary.as_ref() {
-                windows.five_hour = Some(to_quota_window(p));
-            }
-        }
-        if windows.seven_day.is_none() {
-            if let Some(s) = limits.secondary.as_ref() {
-                windows.seven_day = Some(to_quota_window(s));
-            }
-        }
-        models.insert("Codex".to_string(), windows);
-    }
-
-    models
-}
-
-fn flatten_models(
-    models_detailed: &BTreeMap<String, ModelWindows>,
-) -> IndexMap<String, QuotaWindow> {
-    let mut models: IndexMap<String, QuotaWindow> = IndexMap::new();
-    for (name, w) in models_detailed {
-        let selected = w
-            .five_hour
-            .clone()
-            .or_else(|| w.seven_day.clone())
-            .or_else(|| w.other.as_ref().and_then(|o| o.first().cloned()));
-        if let Some(qw) = selected {
-            models.insert(name.clone(), qw);
-        }
-    }
-    models
-}
-
-fn pick_primary(
-    limits: &CodexRateLimits,
-    models_detailed: &BTreeMap<String, ModelWindows>,
-) -> Option<QuotaWindow> {
-    if let Some(p) = limits.primary.as_ref() {
-        return Some(to_quota_window(p));
-    }
-    for m in models_detailed.values() {
-        if let Some(fh) = m.five_hour.as_ref() {
-            return Some(fh.clone());
-        }
-    }
-    for m in models_detailed.values() {
-        if let Some(sd) = m.seven_day.as_ref() {
-            return Some(sd.clone());
-        }
-    }
-    None
-}
-
-fn pick_secondary(
-    limits: &CodexRateLimits,
-    models_detailed: &BTreeMap<String, ModelWindows>,
-) -> Option<QuotaWindow> {
-    if let Some(s) = limits.secondary.as_ref() {
-        return Some(to_quota_window(s));
-    }
-    for m in models_detailed.values() {
-        if let Some(sd) = m.seven_day.as_ref() {
-            return Some(sd.clone());
-        }
-    }
-    None
-}
-
-/// CodexRateLimits → ProviderQuota. `error` embutido se sem janelas usáveis.
-pub fn build_codex_quota(limits: &CodexRateLimits, base: ProviderQuota) -> ProviderQuota {
-    let models_detailed = build_model_windows(limits);
-    let models = flatten_models(&models_detailed);
-    let primary = pick_primary(limits, &models_detailed);
-    let secondary = pick_secondary(limits, &models_detailed);
-
-    if primary.is_none() && secondary.is_none() && models_detailed.is_empty() {
-        return ProviderQuota {
-            error: Some(crate::providers::error::CodexError::NoQuotaWindows.to_string()),
-            ..base
-        };
-    }
-
-    // Credits → extraUsage.
-    let credits_extra: Option<ExtraUsage> = limits.credits.as_ref().and_then(|c| {
-        let balance: f64 = c.balance.parse().unwrap_or(0.0);
-        if c.has_credits || balance > 0.0 {
-            Some(ExtraUsage {
-                enabled: true,
-                remaining: if c.unlimited {
-                    100.0
-                } else {
-                    100.0_f64.min(balance.round())
-                },
-                limit: if c.unlimited { -1.0 } else { 0.0 },
-                used: 0.0,
-            })
-        } else {
-            None
-        }
-    });
-
-    let extra = if !models_detailed.is_empty() || credits_extra.is_some() {
-        Some(ProviderExtra::Codex(CodexQuotaExtra {
-            models_detailed: if models_detailed.is_empty() {
-                None
-            } else {
-                Some(models_detailed)
-            },
-            extra_usage: credits_extra,
-        }))
-    } else {
-        None
-    };
-
-    let plan = normalize_plan(limits.plan_type.as_deref());
-
-    ProviderQuota {
-        available: true,
-        primary,
-        secondary,
-        models: if models.is_empty() {
-            None
-        } else {
-            Some(models)
-        },
-        plan_type: limits.plan_type.clone().filter(|s| !s.is_empty()),
-        plan,
-        extra,
-        ..base
-    }
-}
-
-// ---- Tipos do app-server (camelCase, Deserialize) ----
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexAppServerWindow {
-    pub used_percent: f64,
-    #[serde(default)]
-    pub window_duration_mins: Option<i64>,
-    #[serde(default)]
-    pub resets_at: Option<i64>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexAppServerLimitBucket {
-    #[serde(default)]
-    pub limit_id: Option<String>,
-    #[serde(default)]
-    pub limit_name: Option<String>,
-    #[serde(default)]
-    pub primary: Option<CodexAppServerWindow>,
-    #[serde(default)]
-    pub secondary: Option<CodexAppServerWindow>,
-    #[serde(default)]
-    pub plan_type: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexAppServerCredits {
-    #[serde(default)]
-    pub has_credits: bool,
-    #[serde(default)]
-    pub unlimited: bool,
-    #[serde(default)]
-    pub balance: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexAppServerRateLimitsReadResult {
-    #[serde(default)]
-    pub rate_limits: Option<CodexAppServerLimitBucket>,
-    #[serde(default)]
-    pub rate_limits_by_limit_id: Option<IndexMap<String, CodexAppServerLimitBucket>>,
-    #[serde(default)]
-    pub credits: Option<CodexAppServerCredits>,
-    #[serde(default)]
-    pub plan_type: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexAppServerAccount {
-    #[serde(default)]
-    pub plan_type: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Deserialize)]
-pub struct CodexAppServerAccountReadResult {
-    #[serde(default)]
-    pub account: Option<CodexAppServerAccount>,
-}
-
-// ---- Normalização app-server → CodexRateLimits ----
-
-fn to_raw_window(raw: &CodexAppServerWindow, fallback_minutes: i64) -> CodexWindowRaw {
-    CodexWindowRaw {
-        used_percent: raw.used_percent,
-        window_minutes: raw.window_duration_mins.unwrap_or(fallback_minutes),
-        resets_at: raw.resets_at.unwrap_or(0),
-    }
-}
-
-fn normalize_bucket(
-    raw: &CodexAppServerLimitBucket,
-    fallback_id: Option<&str>,
-) -> Option<CodexLimitBucket> {
-    let limit_id = raw
-        .limit_id
-        .clone()
-        .or_else(|| fallback_id.map(str::to_string))?;
-    let primary = raw.primary.as_ref().map(|w| to_raw_window(w, 300));
-    let secondary = raw.secondary.as_ref().map(|w| to_raw_window(w, 10080));
-    if primary.is_none() && secondary.is_none() {
-        return None;
-    }
-    Some(CodexLimitBucket {
-        limit_id,
-        limit_name: raw.limit_name.clone(),
-        primary,
-        secondary,
-    })
-}
-
-pub fn normalize_appserver_rate_limits(
-    raw: &CodexAppServerRateLimitsReadResult,
-    account_plan_type: Option<&str>,
-) -> Option<CodexRateLimits> {
-    let mut buckets: IndexMap<String, CodexLimitBucket> = IndexMap::new();
-    if let Some(by_id) = raw.rate_limits_by_limit_id.as_ref() {
-        for (limit_id, bucket) in by_id {
-            if let Some(n) = normalize_bucket(bucket, Some(limit_id)) {
-                buckets.insert(n.limit_id.clone(), n);
-            }
-        }
-    }
-    let root = raw.rate_limits.as_ref();
-    let root_bucket = root.and_then(|r| {
-        let fallback = r.limit_id.as_deref().unwrap_or("codex");
-        normalize_bucket(r, Some(fallback))
-    });
-    if let Some(rb) = root_bucket.as_ref() {
-        if !buckets.contains_key(&rb.limit_id) {
-            buckets.insert(rb.limit_id.clone(), rb.clone());
-        }
-    }
-
-    let first = buckets.values().next();
-    let primary = root_bucket
-        .as_ref()
-        .and_then(|b| b.primary.clone())
-        .or_else(|| first.and_then(|b| b.primary.clone()));
-    let secondary = root_bucket
-        .as_ref()
-        .and_then(|b| b.secondary.clone())
-        .or_else(|| first.and_then(|b| b.secondary.clone()));
-
-    if primary.is_none() && secondary.is_none() && buckets.is_empty() {
-        return None;
-    }
-
-    let credits = raw.credits.as_ref().map(|c| CodexCredits {
-        has_credits: c.has_credits,
-        unlimited: c.unlimited,
-        balance: c.balance.clone().unwrap_or_else(|| "0".to_string()),
-    });
-
-    let plan_type = account_plan_type
-        .map(str::to_string)
-        .or_else(|| raw.plan_type.clone())
-        .or_else(|| root.and_then(|r| r.plan_type.clone()));
-
-    Some(CodexRateLimits {
-        primary,
-        secondary,
-        credits,
-        plan_type,
-        buckets: if buckets.is_empty() {
-            None
-        } else {
-            Some(buckets)
-        },
-    })
-}
-
-// ---- Protocolo app-server (JSON-RPC sobre AsyncRead/AsyncWrite) ----
-
-use tokio::io::{AsyncBufReadExt, AsyncWrite, AsyncWriteExt, BufReader};
-
-#[derive(serde::Deserialize)]
-struct AppServerResponse {
-    #[serde(default)]
-    id: Option<i64>,
-    #[serde(default)]
-    result: Option<serde_json::Value>,
-    #[serde(default)]
-    error: Option<serde_json::Value>,
-}
-
-async fn write_json<W: AsyncWrite + Unpin>(
-    w: &mut W,
-    v: &serde_json::Value,
-) -> std::io::Result<()> {
-    let mut s = serde_json::to_string(v).unwrap_or_default();
-    s.push('\n');
-    w.write_all(s.as_bytes()).await
-}
-
-/// Roda o handshake JSON-RPC sobre `reader`/`writer` genéricos e retorna
-/// `Some(CodexRateLimits)` em caso de sucesso ou `None` em timeout/EOF/erro.
-/// Port de `fetchRateLimitsViaAppServer` (codex.ts ~359-453).
-pub async fn run_appserver_protocol<R, W>(
-    reader: R,
-    mut writer: W,
-    version: &str,
-    timeout: std::time::Duration,
-) -> Option<CodexRateLimits>
-where
-    R: tokio::io::AsyncRead + Unpin,
-    W: AsyncWrite + Unpin,
-{
-    use crate::app_identity::APP_NAME;
-
-    let init = serde_json::json!({
-        "method": "initialize",
-        "id": 0,
-        "params": {
-            "clientInfo": {
-                "name": APP_NAME,
-                "title": APP_NAME,
-                "version": version
-            }
-        }
-    });
-    write_json(&mut writer, &init).await.ok()?;
-
-    let mut lines = BufReader::new(reader).lines();
-    // None = não recebido ainda; Some(None) = recebido mas sem plan_type
-    let mut account_plan: Option<Option<String>> = None;
-    let mut rate_limits: Option<CodexAppServerRateLimitsReadResult> = None;
-
-    let hard = tokio::time::sleep(timeout);
-    tokio::pin!(hard);
-    // grace começa longe no futuro (timeout + 1s) para nunca disparar antes de ser armado.
-    let grace = tokio::time::sleep(timeout + std::time::Duration::from_secs(1));
-    tokio::pin!(grace);
-    let mut grace_armed = false;
-
-    loop {
-        tokio::select! {
-            _ = &mut hard => return None,
-            _ = &mut grace, if grace_armed => {
-                return rate_limits.as_ref().and_then(|r| {
-                    let plan = account_plan.as_ref().and_then(|o| o.as_deref());
-                    normalize_appserver_rate_limits(r, plan)
-                });
-            }
-            line = lines.next_line() => {
-                let line = match line {
-                    Ok(Some(l)) => l,
-                    _ => return None,
-                };
-                let msg: AppServerResponse = match serde_json::from_str(&line) {
-                    Ok(m) => m,
-                    Err(_) => continue,
-                };
-                match msg.id {
-                    Some(0) if msg.result.is_some() => {
-                        let _ = write_json(
-                            &mut writer,
-                            &serde_json::json!({"method": "initialized", "params": {}}),
-                        )
-                        .await;
-                        let _ = write_json(
-                            &mut writer,
-                            &serde_json::json!({
-                                "method": "account/read",
-                                "id": 1,
-                                "params": {"refreshToken": false}
-                            }),
-                        )
-                        .await;
-                        let _ = write_json(
-                            &mut writer,
-                            &serde_json::json!({
-                                "method": "account/rateLimits/read",
-                                "id": 2,
-                                "params": {}
-                            }),
-                        )
-                        .await;
-                    }
-                    Some(1) => {
-                        let plan = msg
-                            .result
-                            .as_ref()
-                            .and_then(|v| {
-                                serde_json::from_value::<CodexAppServerAccountReadResult>(
-                                    v.clone(),
-                                )
-                                .ok()
-                            })
-                            .and_then(|a| a.account)
-                            .and_then(|a| a.plan_type);
-                        account_plan = Some(plan);
-                        if let Some(r) = rate_limits.as_ref() {
-                            let plan_ref =
-                                account_plan.as_ref().and_then(|o| o.as_deref());
-                            return normalize_appserver_rate_limits(r, plan_ref);
-                        }
-                    }
-                    Some(2) => {
-                        if let Some(err) = msg.error.as_ref() {
-                            // app-server respondeu com erro pro rate-limits (ex.: token
-                            // expirado/revogado → 401 no backend). Não virá outra
-                            // resposta pro id=2; sair já em vez de esperar o hard
-                            // timeout ocioso.
-                            log::debug!("Codex app-server rateLimits/read error: {err}");
-                            return None;
-                        }
-                        let parsed = msg.result.as_ref().and_then(|v| {
-                            serde_json::from_value::<CodexAppServerRateLimitsReadResult>(
-                                v.clone(),
-                            )
-                            .ok()
-                        });
-                        if let Some(r) = parsed {
-                            let has_data =
-                                r.rate_limits.is_some() || r.rate_limits_by_limit_id.is_some();
-                            if has_data {
-                                rate_limits = Some(r);
-                                if account_plan.is_some() {
-                                    if let Some(rr) = rate_limits.as_ref() {
-                                        let plan_ref = account_plan
-                                            .as_ref()
-                                            .and_then(|o| o.as_deref());
-                                        return normalize_appserver_rate_limits(rr, plan_ref);
-                                    }
-                                } else {
-                                    grace.as_mut().reset(
-                                        tokio::time::Instant::now()
-                                            + std::time::Duration::from_millis(200),
-                                    );
-                                    grace_armed = true;
-                                }
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-}
-
-// ---- Fallback session-log ----
-
-/// Acha o `.jsonl` mais recente em `sessions_dir/YYYY/MM/DD` (hoje ou ontem, hora local).
-pub fn find_latest_session_file(
-    sessions_dir: &Path,
-    now_ms: u64,
-    offset: UtcOffset,
-) -> Option<PathBuf> {
-    let now = OffsetDateTime::from_unix_timestamp_nanos((now_ms as i128) * 1_000_000)
-        .ok()?
-        .to_offset(offset);
-    for day_offset in 0..2i64 {
-        let date = now - TimeDuration::days(day_offset);
-        let dir = sessions_dir
-            .join(format!("{:04}", date.year()))
-            .join(format!("{:02}", date.month() as u8))
-            .join(format!("{:02}", date.day()));
-        let mut newest: Option<(std::time::SystemTime, PathBuf)> = None;
-        let entries = match std::fs::read_dir(&dir) {
-            Ok(e) => e,
-            Err(_) => continue,
-        };
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.extension().and_then(|e| e.to_str()) != Some("jsonl") {
-                continue;
-            }
-            let mtime = match entry.metadata().and_then(|m| m.modified()) {
-                Ok(t) => t,
-                Err(_) => continue,
-            };
-            if newest.as_ref().map(|(t, _)| mtime > *t).unwrap_or(true) {
-                newest = Some((mtime, path));
-            }
-        }
-        if let Some((_, path)) = newest {
-            return Some(path);
-        }
-    }
-    None
-}
-
-#[derive(Deserialize)]
-struct SessionEvent {
-    #[serde(default)]
-    payload: Option<SessionPayload>,
-}
-
-#[derive(Deserialize)]
-struct SessionPayload {
-    #[serde(rename = "type", default)]
-    kind: Option<String>,
-    #[serde(default)]
-    rate_limits: Option<CodexRateLimits>,
-}
-
-/// Lê o arquivo de sessão, faz scan reverso e retorna o primeiro `token_count` com `rate_limits`.
-pub fn extract_rate_limits(path: &Path) -> Option<CodexRateLimits> {
-    let content = std::fs::read_to_string(path).ok()?;
-    for line in content.trim().lines().rev() {
-        if line.trim().is_empty() {
-            continue;
-        }
-        if let Ok(ev) = serde_json::from_str::<SessionEvent>(line) {
-            if let Some(p) = ev.payload {
-                if p.kind.as_deref() == Some("token_count") {
-                    if let Some(rl) = p.rate_limits {
-                        return Some(rl);
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-// ---- Provider (QuotaSource + Provider impls) ----
-
-use std::process::Stdio;
+use async_trait::async_trait;
 
 use super::base::{base_get_quota, QuotaSource};
 use super::error::{CodexError, ProviderError};
+use super::types::ProviderQuota;
 use super::{Ctx, Provider};
-use async_trait::async_trait;
+use app_server::fetch_via_appserver;
 
-async fn fetch_via_appserver(version: &str) -> Option<CodexRateLimits> {
-    let mut child = tokio::process::Command::new("codex")
-        .arg("app-server")
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::null())
-        .kill_on_drop(true)
-        .spawn()
-        .ok()?;
-    let stdout = child.stdout.take()?;
-    let stdin = child.stdin.take()?;
-    let result =
-        run_appserver_protocol(stdout, stdin, version, std::time::Duration::from_secs(4)).await;
-    let _ = child.start_kill();
-    result
-}
+// ---- Provider (QuotaSource + Provider impls) ----
 
 pub struct CodexProvider;
 
@@ -804,9 +103,14 @@ impl Provider for CodexProvider {
     }
 }
 
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use super::super::iso_from_ms;
+    use super::super::types::{CodexQuotaExtra, ProviderExtra, ProviderQuota};
+    use indexmap::IndexMap;
+    use time::UtcOffset;
 
     fn base() -> ProviderQuota {
         ProviderQuota {
